+++ actix-web-lab/examples/sse.rs
@@ -2,18 +2,15 @@
 
 use std::{convert::Infallible, io, time::Duration};
 
-use actix_web::{get, middleware::Logger, App, HttpRequest, HttpServer, Responder};
-#[allow(deprecated)]
-use actix_web_lab::respond::Html;
+use actix_web::{get, middleware::Logger, web::Html, App, HttpRequest, HttpServer, Responder};
 use actix_web_lab::{extract::Path, sse};
 use futures_util::stream;
 use time::format_description::well_known::Rfc3339;
 use tokio::time::sleep;
 
-#[allow(deprecated)]
 #[get("/")]
 async fn index() -> impl Responder {
-    Html(include_str!("./assets/sse.html").to_string())
+    Html::new(include_str!("./assets/sse.html").to_string())
 }
 
 /// Countdown event stream starting from 8.